Removed static reference in visibility.c and connected it to the struct cdev.
Removed static reference in visibility.c and connected it to the struct cdev.
This is in preparation of moving plugin code out side of the simulator when
it is moved to /sys/dev. Plugin code maybe in /src/tools/tools?

Need to think about what happens in we unload a plugin device but hal still
uses it, should we reject unloading? Or do we deregister ourself and stop
simulation? Maybe vert to a "basic" plugin?